abstract Connectors to component
